+++ lib/src/builders/builder.dart
@@ -3,14 +3,16 @@ import 'package:flutter/material.dart';
 
 import '../ast.dart';
 import '../helpers/inline_wraper.dart';
+import '../helpers/is_dark_mode.dart';
 import '../models/markdown_tree_element.dart';
 import '../renderer.dart';
 
 abstract class MarkdownElementBuilder {
   MarkdownElementBuilder({
+    this.context,
     this.textStyle,
     this.textStyleMap,
-  }) {
+  }) : darkMode = isDarkMode(context) {
     assert(matchTypes.isNotEmpty);
     assert(
       textStyle == null || textStyleMap == null,
@@ -35,6 +37,10 @@ abstract class MarkdownElementBuilder {
 
   /// [TextStyle] of the current element.
   final TextStyle? textStyle;
+
+  final BuildContext? context;
+
+  final bool darkMode;
 
   /// If [matchTypes] matches more than one type, you need to use textStyleMap
   /// to set [TextStyle] for each type.