+++ lib/src/renderer.dart
@@ -26,6 +26,7 @@ import 'transformer.dart';
 
 class MarkdownRenderer implements NodeVisitor {
   MarkdownRenderer({
+    BuildContext? context,
     required MarkdownStyle styleSheet,
     MarkdownTapLinkCallback? onTapLink,
     MarkdownListItemMarkerBuilder? listItemMarkerBuilder,
@@ -42,6 +43,14 @@ class MarkdownRenderer implements NodeVisitor {
         _selectionRegistrar = selectionRegistrar,
         _blockSpacing = styleSheet.blockSpacing,
         _styleSheet = styleSheet,
+        _defaultTextStyle = TextStyle(
+          fontSize: 16,
+          height: 1.5,
+          color: (context != null
+                  ? Theme.of(context).textTheme.bodyText2?.color
+                  : null) ??
+              const Color(0xff333333),
+        ),
         _textAlign = textAlign ?? TextAlign.start {
     final defaultBuilders = [
       HeadlineBuilder(
@@ -59,6 +68,7 @@ class MarkdownRenderer implements NodeVisitor {
         h6Padding: styleSheet.h6Padding,
       ),
       SimpleInlinesBuilder(
+        context: context,
         emphasis: styleSheet.emphasis,
         strongEmphasis: styleSheet.strongEmphasis,
         highlight: styleSheet.highlight,
@@ -76,7 +86,7 @@ class MarkdownRenderer implements NodeVisitor {
         textStyle: styleSheet.paragraph,
         padding: styleSheet.paragraphPadding,
       ),
-      CodeSpanBuilder(textStyle: styleSheet.codeSpan),
+      CodeSpanBuilder(context: context, textStyle: styleSheet.codeSpan),
       LinkBuilder(textStyle: styleSheet.link, onTap: onTapLink),
       TableBuilder(
         table: styleSheet.table,
@@ -93,6 +103,7 @@ class MarkdownRenderer implements NodeVisitor {
         enableImageSize: enableImageSize,
       ),
       CodeBlockBuilder(
+        context: context,
         textStyle: styleSheet.codeBlock,
         padding: styleSheet.codeblockPadding,
         decoration: styleSheet.codeblockDecoration,
@@ -101,6 +112,7 @@ class MarkdownRenderer implements NodeVisitor {
         copyIconColor: styleSheet.copyIconColor,
       ),
       BlockquoteBuilder(
+        context: context,
         textStyle: styleSheet.blockquote,
         decoration: styleSheet.blockquoteDecoration,
         padding: styleSheet.blockquotePadding,
@@ -137,6 +149,7 @@ class MarkdownRenderer implements NodeVisitor {
   final Color? _selectionColor;
   final SelectionRegistrar? _selectionRegistrar;
   final MarkdownStyle _styleSheet;
+  final TextStyle _defaultTextStyle;
 
   bool get selectable => _selectionColor != null && _selectionRegistrar != null;
   MouseCursor? get mouseCursor => selectable ? SystemMouseCursors.text : null;
@@ -184,11 +197,7 @@ class MarkdownRenderer implements NodeVisitor {
       builder.gestureRecognizer(element),
     );
 
-    final defaultTextStyle = const TextStyle(
-      fontSize: 16,
-      height: 1.5,
-      color: Color(0xff333333),
-    ).merge(_styleSheet.textStyle);
+    final defaultTextStyle = _defaultTextStyle.merge(_styleSheet.textStyle);
 
     _tree.add(_TreeElement.fromAstElement(
       element,